+++ subtest-impl/tests/expand_impl/unused_mut_in_subtest_parameter.rs
@@ -1,0 +1,12 @@
+#[subtest]
+#[rstest]
+#[case(1)]
+fn value_can_be_incremented(#[case] mut value: u32) {
+    #[subtest]
+    fn value_is_positive() {
+        assert!(value > 0);
+    }
+
+    value += 1;
+    assert!(value > 1);
+}