+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next Release
 
+## v1.0.1 (2026-08-22)
+
 ### Fixed
 
 * `mut` on variable and parameter declarations is no longer reported as unused if the variable or parameter is not modified in a nested subtest, but modified in the parent test function. Limitation: Does not work in a crate which sets `#![forbid(unused_mut)]`, as `forbid` rejects the generated `#[allow]` attributes